Fix a bunch of performance issues
Looked through the profiler output, noticed a bunch of time was
wasted in property lookups so changed a bunch of things to direct
member access.  Also stopped using gestouch to detect taps, as
its overkill and the delay to ensure its a tap as opposed to
some other gesture is unacceptable for a quick reaction game.
Also minor art tweak here.